RCJ, Removed the files common_utils.h/.cpp and tested the code.
