+++ src/catheter_arduino_gui/inc/com/pc_utils.h
@@ -8,7 +8,7 @@
 #ifndef PC_UTILS_H
 #define PC_UTILS_H
 
-#include "common_utils.h"
+#include "com/catheter_commands.h"
 
 /** \brief int loadPlayFile(const char*,std::vector<CatheterChannelCmd> &): Loads and parses a playfile.
  * The playfile is a recorded vector of computed commands with inter command delay. */
